+++ core/server/public/ghost-sdk.js
